+++ librechat.example.yaml
@@ -2,7 +2,7 @@
 # https://docs.librechat.ai/install/configuration/custom_config.html
 
 # Configuration version (required)
-version: 1.0.5
+version: 1.0.6
 
 # Cache settings: Set to true to enable caching
 cache: true
